# Conflicts:
#	src/main/java/com/team1/efep/service_implementors/AccountServiceImpl.java

+++ src/main/java/com/team1/efep/service_implementors/BuyerServiceImpl.java
@@ -12,7 +12,6 @@ import com.team1.efep.models.request_models.*;
 import com.team1.efep.models.response_models.*;
 import com.team1.efep.repositories.*;
 import com.team1.efep.services.BuyerService;
-import com.team1.efep.utils.ConvertMapIntoStringUtil;
 import com.team1.efep.utils.FileReaderUtil;
 import com.team1.efep.utils.OTPGeneratorUtil;
 import com.team1.efep.utils.OutputCheckerUtil;
@@ -585,7 +584,7 @@ public class BuyerServiceImpl implements BuyerService {
         Account account = Role.getCurrentLoggedAccount(request.getAccountId(), accountRepo);
         Order order = orderRepo.findById(request.getOrderId()).orElse(null);
         assert order != null;
-        Map<String, String> error = ViewOrderDetailValidation.validate(request, account, order);
+        Map<String, String> error = ViewOrderDetailValidation.validate(request, order);
         if (!error.isEmpty()) {
             return error;
         }
@@ -1049,6 +1048,8 @@ public class BuyerServiceImpl implements BuyerService {
         Map<Seller, List<WishlistItem>> itemsBySeller = items.stream()
                 .collect(Collectors.groupingBy(item -> item.getFlower().getSeller()));
 
+        boolean sendEmailResult = false;
+
         for (Map.Entry<Seller, List<WishlistItem>> entry : itemsBySeller.entrySet()) {
             Seller seller = entry.getKey();
             List<WishlistItem> sellerItems = entry.getValue();
@@ -1067,7 +1068,8 @@ public class BuyerServiceImpl implements BuyerService {
                     .status(Status.ORDER_STATUS_PROCESSING)
                     .paymentMethod(paymentMethod)
                     .build());
-            sendOrderEmail(savedOrder, user);
+
+            List<OrderDetail> orderDetails = new ArrayList<>();
             for (WishlistItem item : entry.getValue()) {
                 Flower flower = item.getFlower();
                 flower.setSoldQuantity(flower.getSoldQuantity() + item.getQuantity());
@@ -1081,15 +1083,22 @@ public class BuyerServiceImpl implements BuyerService {
                         .price(item.getFlower().getPrice())
                         .build();
                 orderDetailRepo.save(orderDetail);
+                orderDetails.add(orderDetail);
                 flowerRepo.save(flower);
             }
+            savedOrder.setOrderDetailList(orderDetails);
+            orderRepo.save(savedOrder);
+            sendEmailResult = sendOrderEmail(savedOrder, user);
+            sendEmailResult = sendOrderSellerEmail(savedOrder, seller);
         }
-        wishlistItemRepo.deleteAll(items);
-        user.getWishlist().setWishlistItemList(new ArrayList<>());
-        userRepo.save(user);
+        if(sendEmailResult) {
+            wishlistItemRepo.deleteAll(items);
+            user.getWishlist().setWishlistItemList(new ArrayList<>());
+            userRepo.save(user);
+        }
     }
 
-    private void sendOrderEmail(Order order, User user) {
+    private boolean sendOrderEmail(Order order, User user) {
 
         MimeMessage message = mailSender.createMimeMessage();
 
@@ -1101,7 +1110,7 @@ public class BuyerServiceImpl implements BuyerService {
 
             helper.setTo(user.getAccount().getEmail());
 
-            helper.setSubject(Const.BUSINESS_PLAN_SUBJECT);
+            helper.setSubject(Const.EMAIL_SUBJECT_ORDER);
 
             String emailContent = FileReaderUtil.readFile(order, user);
 
@@ -1110,8 +1119,35 @@ public class BuyerServiceImpl implements BuyerService {
             mailSender.send(message);
 
         } catch (MessagingException e) {
-            throw new RuntimeException(e);
+            return false;
         }
+        return true;
+    }
+
+    private boolean sendOrderSellerEmail(Order order, Seller seller) {
+
+        MimeMessage message = mailSender.createMimeMessage();
+
+        MimeMessageHelper helper = null;
+        try {
+            helper = new MimeMessageHelper(message, true, "UTF-8");
+
+            helper.setFrom("vannhuquynhp@gmail.com");
+
+            helper.setTo(seller.getUser().getAccount().getEmail());
+
+            helper.setSubject(Const.EMAIL_SUBJECT_ORDER);
+
+            String emailContent = FileReaderUtil.readFile(order);
+
+            helper.setText(emailContent, true);
+
+            mailSender.send(message);
+
+        } catch (MessagingException e) {
+            return false;
+        }
+        return true;
     }
 
     //-----------------------------------GET COD PAYMENT RESULT--------------------------------------//
@@ -1234,19 +1270,6 @@ public class BuyerServiceImpl implements BuyerService {
         model.addAttribute("error", (Map<String, String>) output);
         return "paymentFailed";
     }
-
-//    @Override
-//    public VNPayResponse getPaymentResultForBuyNowAPI(Map<String, String> params, int accountId, HttpServletRequest httpServletRequest) {
-//
-//        Object output = getPaymentResultForBuyNowLogic(params, accountId, httpServletRequest);
-//        if (OutputCheckerUtil.checkIfThisIsAResponseObject(output, VNPayResponse.class)) {
-//            return (VNPayResponse) output;
-//        }
-//        return VNPayResponse.builder()
-//                .status("400")
-//                .message(ConvertMapIntoStringUtil.convert((Map<String, String>) output))
-//                .build();
-//    }
 
     private Object getPaymentResultForBuyNowLogic(Map<String, String> params, int flowerId, int quantity, int accountId, HttpServletRequest httpServletRequest) {
         User user = Role.getCurrentLoggedAccount(accountId, accountRepo).getUser();
